--- .coverage
@@ -1,1 +1,0 @@
-!coverage.py: This is a private format, don't read it directly!{"lines":{"C:\\Users\\mk\\Desktop\\Andela_Bootcamp\\i_reporter_api\\tests\\__init__.py":[1],"C:\\Users\\mk\\Desktop\\Andela_Bootcamp\\i_reporter_api\\tests\\test_auth.py":[1,2,3,4,6,7,11,17,25,33,41,50,59,68,77,84,89,97,106,114,123,132,8,9,85,86,87,90,91,92,93,94,95,98,99,100,102,103,104,115,116,117,118,119,120,121,107,108,109,110,111,112,133,134,135,136,137,138,139,124,125,126,127,128,129,130,78,79,80,81,82,12,13,14,15,26,27,28,29,30,31,34,35,36,37,38,39,69,70,71,72,73,74,75,60,61,62,63,64,65,66,51,52,53,54,55,56,57,18,19,20,21,22,23,42,43,44,45,46,47,48],"C:\\Users\\mk\\Desktop\\Andela_Bootcamp\\i_reporter_api\\app\\__init__.py":[1],"C:\\Users\\mk\\Desktop\\Andela_Bootcamp\\i_reporter_api\\app\\views.py":[1,2,3,4,5,6,7,10,12,13,15,16,17,19,31,48,66,105,123,143,160,161,162,68,69,70,71,72,73,74,75,76,77,78,79,80,82,83,85,86,88,90,92,93,94,96,100,98,102,146,147,148,149,150,151,155,157,153,50,51,52,53,54,55,56,60,62,58,33,34,35,36,37,38,39,43,45,41,21,22,23,24,25,26,27,28,29,126,127,128,129,131,132,133,134,138,136,140,107,108,109,110,111,112,113,114,118,116,120],"C:\\Users\\mk\\Desktop\\Andela_Bootcamp\\i_reporter_api\\instance\\__init__.py":[1],"C:\\Users\\mk\\Desktop\\Andela_Bootcamp\\i_reporter_api\\instance\\config.py":[1,3,4,5,6,7,9,10,11,12,14,15,16,17,19,20,21,23,24,25,26,29,30,31,32],"C:\\Users\\mk\\Desktop\\Andela_Bootcamp\\i_reporter_api\\app\\data_store\\__init__.py":[1],"C:\\Users\\mk\\Desktop\\Andela_Bootcamp\\i_reporter_api\\app\\data_store\\data.py":[1,3,8,10,11,15,16,17,19,20,21,22,24,27,28,29,30,31,32,34,35,36,37,38,39,40,41,42,43,46,47,48,49,50,52,54,56,60],"C:\\Users\\mk\\Desktop\\Andela_Bootcamp\\i_reporter_api\\app\\models\\__init__.py":[1],"C:\\Users\\mk\\Desktop\\Andela_Bootcamp\\i_reporter_api\\app\\models\\user.py":[1,2,3,4,5,7,8,20,23,26,44,48,54,9,10,11,12,13,14,15,16,17,18,46,27,29,30,31,32,34,35,36,37,39,21,56,57,58,59,60],"C:\\Users\\mk\\Desktop\\Andela_Bootcamp\\i_reporter_api\\app\\models\\incident.py":[1,2,5,6,18,8,9,10,11,12,13,14,15,16,19],"C:\\Users\\mk\\Desktop\\Andela_Bootcamp\\i_reporter_api\\app\\utilities\\helper_functions.py":[1,2,3,4,7,9,13,21,29,35,42,54,66,73,80,88,95,97,98,99,11,90,91,92,93,15,16,18,19,45,46,47,48,49,50,51,37,40,38,39,17,82,83,84,85,86,31,32,33,23,24,25,26,27,75,76,77,78,68,69,70,71],"C:\\Users\\mk\\Desktop\\Andela_Bootcamp\\i_reporter_api\\app\\utilities\\incident_validation.py":[4,5,6,7,8,9,10,13,14,23,15,16,17,18,19,20,21,24,26,28,30,32,34,36,38,37,25,27,33,29,31,35],"C:\\Users\\mk\\Desktop\\Andela_Bootcamp\\i_reporter_api\\tests\\test_incident.py":[1,2,3,4,6,7,23,28,35,51,58,65,72,79,89,99,109,119,129,141,160,168,188,197,205,212,222,234,244,260,267,273,282,301,321,337,344,350,359,378,8,9,11,12,14,15,16,17,18,20,52,53,54,55,56,73,74,75,76,77,345,346,347,348,130,131,132,133,134,135,136,137,138,139,80,81,82,83,84,85,86,87,100,101,102,103,104,105,106,107,213,214,215,216,217,218,219,220,120,121,122,123,124,125,126,127,235,236,237,238,239,240,241,242,268,269,270,271,223,224,225,226,227,228,229,230,231,232,110,111,112,113,114,115,116,117,90,91,92,93,94,95,96,97,198,199,200,201,202,203,351,352,353,354,355,356,274,275,276,277,278,279,206,207,208,209,210,59,60,61,62,63,338,339,340,341,342,261,262,263,264,265,66,67,68,69,70,37,38,39,40,41,42,43,45,46,47,48,49,323,324,325,326,327,328,329,331,332,333,334,335,246,247,248,249,250,251,252,254,255,256,257,258,29,30,31,32,33,24,25,26,170,171,172,173,174,175,176,178,179,180,181,182,183,184,185,186,361,362,363,364,365,366,367,369,370,371,372,373,374,375,189,190,191,192,193,194,195,284,285,286,287,288,289,290,292,293,294,295,296,297,298,143,144,145,146,147,148,149,151,152,153,154,155,156,157,158,380,381,382,383,384,385,386,388,389,390,391,392,393,394,161,162,163,164,165,166,303,304,305,306,307,308,309,311,312,313,314,315,316,317],"C:\\Users\\mk\\Desktop\\Andela_Bootcamp\\i_reporter_api\\app\\auth\\__init__.py":[1,3,5],"C:\\Users\\mk\\Desktop\\Andela_Bootcamp\\i_reporter_api\\app\\auth\\views.py":[2,3,4,5,6,7,8,9,10,11,12,14,15,16,46,47,49,87,88,90,91,92,93,95,96,97,98,51,53,54,56,58,60,61,62,63,64,66,67,69,70,72,84,76,74,18,20,21,22,23,24,25,26,28,29,31,33,34,35,37,38,40,41,43],"C:\\Users\\mk\\Desktop\\Andela_Bootcamp\\i_reporter_api\\app\\utilities\\login_validation.py":[4,5,9,10,14,11,12,15,17,19,21,18,20,16],"C:\\Users\\mk\\Desktop\\Andela_Bootcamp\\i_reporter_api\\app\\utilities\\register_validation.py":[1,5,6,7,8,9,10,11,15,17,26,18,19,20,21,22,23,24,27,29,31,33,35,37,39,41,30,32,40,38,36,28,34]}}
