+++ tests/test_incident.py
@@ -240,155 +240,3 @@ class TestFlaskApi(unittest.TestCase):
         self.assertEqual(data["status"], 400)
         self.assertEqual(
             data["error"], "Valid status required. Status should be of type string")
-
-    def test_get_list_of_incidents_no_token_header(self):
-        redflag_incident = {
-                                "created_by": 5242,
-                                "type": "red-flag",
-                                "location": "0.112, 0.545",
-                                "status": " sdfsd",
-                                "images": ["sdfaf", "vfdgdf"],
-                                "videos": ["video link", "fgfdgs"],
-                                "comment": "This is the comment sgfd"
-                            }
-        self.client.post('/api/v1/red-flags', data=json.dumps(redflag_incident),
-                         content_type='application/json', headers=self.headers)
-        response = self.client.get('/api/v1/red-flags')
-        data = json.loads(response.data)
-        self.assertEqual(data["status"], 401)
-
-    def test_get_an_incident_no_token_header(self):
-        id = incidents[0].id
-        response = self.client.get(
-            '/api/v1/red-flags/'+str(id))
-        data = json.loads(response.data)
-        self.assertEqual(data["status"], 401)
-
-    def test_create_new_red_flag_no_token_header(self):
-        response = self.client.post('/api/v1/red-flags', data=json.dumps(incident6_data_dictionary),
-                                    content_type='application/json')
-        data = json.loads(response.data)
-        self.assertEqual(data["status"], 401)
-
-    def test_delete_red_flag_given_id_no_token_header(self):
-        incidents.append(incident3)
-        id = incidents[0].id
-        response = self.client.delete(
-            '/api/v1/red-flags/'+str(id))
-        data = json.loads(response.data)
-        self.assertEqual(data["status"], 401)
-
-
-    def test_update_red_flag_comment_no_token_header(self):
-        redflag_incident = {
-                                "created_by": 5242,
-                                "type": "red-flag",
-                                "location": "0.112, 0.545",
-                                "status": " sdfsd",
-                                "images": ["sdfaf", "vfdgdf"],
-                                "videos": ["video link", "fgfdgs"],
-                                "comment": "This is the comment sgfd"
-                            }
-        self.client.post('/api/v1/red-flags', data=json.dumps(redflag_incident),
-                         content_type='application/json', headers=self.headers)
-        id = incidents[0].id
-        response = self.client.patch('/api/v1/red-flags/'+str(id)+'/comment', data=json.dumps(new_comment),
-                                     content_type='application/json')
-        data = json.loads(response.data)
-        self.assertEqual(data["status"], 401)
-
-
-    def test_update_red_flag_location_no_header_token(self):
-        redflag_incident = {
-                                "created_by": 5242,
-                                "type": "red-flag",
-                                "location": "0.112, 0.545",
-                                "status": " sdfsd",
-                                "images": ["sdfaf", "vfdgdf"],
-                                "videos": ["video link", "fgfdgs"],
-                                "comment": "This is the comment sgfd"
-                            }
-        self.client.post('/api/v1/red-flags', data=json.dumps(redflag_incident),
-                            content_type='application/json', headers=self.headers)
-        id = incidents[0].id
-        response = self.client.patch('/api/v1/red-flags/'+str(id)+'/location', data=json.dumps(new_location),
-                                        content_type='application/json')
-        data = json.loads(response.data)
-        self.assertEqual(data["status"], 401)
-
-
-
-    def test_get_list_of_incidents_expired_token_header(self):
-        redflag_incident = {
-                                "created_by": 5242,
-                                "type": "red-flag",
-                                "location": "0.112, 0.545",
-                                "status": " sdfsd",
-                                "images": ["sdfaf", "vfdgdf"],
-                                "videos": ["video link", "fgfdgs"],
-                                "comment": "This is the comment sgfd"
-                            }
-        self.client.post('/api/v1/red-flags', data=json.dumps(redflag_incident),
-                         content_type='application/json', headers=self.headers)
-        response = self.client.get('/api/v1/red-flags', headers=self.header_old)
-        data = json.loads(response.data)
-        self.assertEqual(data["status"], 401)
-
-    def test_get_an_incident_expired_token_header(self):
-        id = incidents[0].id
-        response = self.client.get(
-            '/api/v1/red-flags/'+str(id), headers=self.header_old)
-        data = json.loads(response.data)
-        self.assertEqual(data["status"], 401)
-
-    def test_create_new_red_flag_expired_token_header(self):
-        response = self.client.post('/api/v1/red-flags', data=json.dumps(incident6_data_dictionary),
-                                    content_type='application/json', headers=self.header_old)
-        data = json.loads(response.data)
-        self.assertEqual(data["status"], 401)
-
-    def test_delete_red_flag_given_id_expired_token_header(self):
-        incidents.append(incident3)
-        id = incidents[0].id
-        response = self.client.delete(
-            '/api/v1/red-flags/'+str(id), headers=self.header_old)
-        data = json.loads(response.data)
-        self.assertEqual(data["status"], 401)
-
-
-    def test_update_red_flag_comment_expired_token_header(self):
-        redflag_incident = {
-                                "created_by": 5242,
-                                "type": "red-flag",
-                                "location": "0.112, 0.545",
-                                "status": " sdfsd",
-                                "images": ["sdfaf", "vfdgdf"],
-                                "videos": ["video link", "fgfdgs"],
-                                "comment": "This is the comment sgfd"
-                            }
-        self.client.post('/api/v1/red-flags', data=json.dumps(redflag_incident),
-                         content_type='application/json', headers=self.headers)
-        id = incidents[0].id
-        response = self.client.patch('/api/v1/red-flags/'+str(id)+'/comment', data=json.dumps(new_comment),
-                                     content_type='application/json', headers=self.header_old)
-        data = json.loads(response.data)
-        self.assertEqual(data["status"], 401)
-
-
-    def test_update_red_flag_location_expired_token_header(self):
-        redflag_incident = {
-                                "created_by": 5242,
-                                "type": "red-flag",
-                                "location": "0.112, 0.545",
-                                "status": " sdfsd",
-                                "images": ["sdfaf", "vfdgdf"],
-                                "videos": ["video link", "fgfdgs"],
-                                "comment": "This is the comment sgfd"
-                            }
-        self.client.post('/api/v1/red-flags', data=json.dumps(redflag_incident),
-                            content_type='application/json', headers=self.headers)
-        id = incidents[0].id
-        response = self.client.patch('/api/v1/red-flags/'+str(id)+'/location', data=json.dumps(new_location),
-                                        content_type='application/json', headers=self.header_old)
-        data = json.loads(response.data)
-        self.assertEqual(data["status"], 401)
